watch-and-learn-main/voyageai-chat/requirements.txt
# VoyageAI Chat Application Dependencies

# Core
voyageai>=0.3.0
openai>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0

# LangChain (prompt management)
langchain>=0.3.0
langchain-openai>=0.3.0

# Testing
pytest>=8.0.0
pytest-benchmark>=4.0.0
